added n choose k + changed signed to unsigne don math algos

+++ tests/Kkommon.Tests/AlgorithmMathTests.cs
@@ -6,14 +6,14 @@ namespace Kkommon.Tests.Math
     {
         [Theory]
         [InlineData(20, 6, 2)][InlineData(4, 20, 4)][InlineData(2, 7, 1)]
-        public void Finds_Greatest_Common_Denominator(int a, int b, int gcd)
+        public void Finds_Greatest_Common_Denominator(ulong a, ulong b, ulong gcd)
         {
             Assert.Equal(Algorithms.Math.Gcd(a, b), gcd);
         }
 
         [Theory]
         [InlineData(20, 6, 60)][InlineData(4, 20, 20)][InlineData(2, 7, 14)]
-        public void Finds_Least_Common_Multiple(int a, int b, int lcm)
+        public void Finds_Least_Common_Multiple(ulong a, ulong b, ulong lcm)
         {
             Assert.Equal(Algorithms.Math.Lcm(a, b), lcm);
         }
